Redux folders renamed as duck.
Initial states moved to reducers.
