+++ src/app/main/MainContainer.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { MainComponent } from './MainComponent';
-import AppState from "../redux/AppState";
+import AppState from "../duck/state";
 
 export const MainContainer = connect(
     ({main : state}: AppState) => ({
