+++ src/app/main/models/Room.ts
@@ -1,4 +1,4 @@
-import Context from "../../../../libraries/Context";
+import Context from "../../../libraries/Context";
 
 export default interface Room extends Context {
 	id? : number,
