+++ src/app/main/rooms/RoomListContainer.tsx
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import {RoomListComponent, RoomListDispatchProps, RoomListStateProps} from "./RoomListComponent";
-import AppState from "../../redux/AppState";
-import {selectRoom} from "../duck/redux/actions";
+import AppState from "../../duck/state";
+import {selectRoom} from "../duck/actions";
 
 export const RoomListContainer = connect(
     ({ main : state }: AppState): RoomListStateProps => ({
