+++ src/app/roomEditor/duck/reducers.ts
@@ -1,7 +1,10 @@
 import {combineReducers} from "redux";
+import RoomEditorState from "./state";
 
 const roomEditorReducers = combineReducers({
 	// sub-stores
 });
+
+export const initialRoomEditorState : RoomEditorState = {}
 
 export default roomEditorReducers;
