+++ src/app/roomEditor/duck/state.ts
@@ -1,0 +1,4 @@
+export default interface RoomEditorState {
+
+}
+
