--- src/app/roomEditor/redux/RoomEditorState.ts
@@ -1,5 +1,0 @@
-export default interface RoomEditorState {
-
-}
-
-export const initialRoomEditorState : RoomEditorState = {}
