+++ src/app/templateEditor/duck/reducers.ts
@@ -1,7 +1,10 @@
 import {combineReducers} from "redux";
+import TemplateEditorState from "./state";
 
 const templateEditorReducers = combineReducers({
 
 });
+
+export const initialTemplateEditorState : TemplateEditorState = {};
 
 export default templateEditorReducers;
