+++ src/app/templateEditor/duck/state.ts
@@ -1,0 +1,3 @@
+export default interface TemplateEditorState {
+
+}
