--- src/app/templateEditor/redux/TemplateEditorState.ts
@@ -1,5 +1,0 @@
-export default interface TemplateEditorState {
-
-}
-
-export const initialTemplateEditorState : TemplateEditorState = {};
